fix(topk): rename flashinfer/top_k package to topk_varlen (avoid shadowing top_k())
The flashinfer/top_k/ package introduced by the relocation shadowed the
existing flashinfer.top_k sampling function (from flashinfer/topk.py):
importing the submodule rebinds flashinfer.top_k to the module, so every
flashinfer.top_k(logits, k) call raised "TypeError: 'module' object is not
callable" -- 429 failures in tests/utils/test_topk.py (the CI JIT Unittest
sampling shard).

Rename the package and its tests dir to the collision-free topk_varlen,
mirroring the top_k_varlen API name. Note top_k_varlen (underscored) would
collide identically with the top_k_varlen function, so topk_varlen is used.

  flashinfer/top_k/  -> flashinfer/topk_varlen/
  tests/top_k/       -> tests/topk_varlen/

Verified: flashinfer.top_k is a callable function again; tests/utils/test_topk.py
962 passed (was 429 failed); tests/topk_varlen/test_topk_varlen.py 115 passed
on sm_100; trace-registry inventory green; pre-commit clean.
