+++ flashinfer/topk_varlen/__init__.py
@@ -15,7 +15,7 @@
 """Top-K decode kernels for FlashInfer.
 
 The CuTe-DSL kernel source (GVR and radix Top-K for Blackwell sm_100+) lives in
-``flashinfer.top_k.kernels``.  The public ``top_k_varlen`` API is defined in
-``flashinfer.top_k.topk_varlen`` and re-exported from the top-level
+``flashinfer.topk_varlen.kernels``.  The public ``top_k_varlen`` API is defined in
+``flashinfer.topk_varlen.topk_varlen`` and re-exported from the top-level
 ``flashinfer`` namespace.
 """
